# specs/001-wiki-platform/contracts/api.yaml
openapi: 3.0.3
info:
  title: Wiki Platform API
  description: API for the wiki platform with article management, user profiles, and social features
  version: 1.0.0
  contact:
    name: Wiki Platform Support
    email: support@wikiplatform.example.com
servers:
  - url: https://api.wikipatform.example.com/v1
    description: Production server
  - url: https://staging-api.wikipatform.example.com/v1
    description: Staging server
paths:
...
  /auth/login:
    post:
      summary: User login
      description: Authenticate user and return access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      summary: User logout
      description: Invalidate the current user session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized
...
  /comments:
    post:
      summary: Create comment
      description: Create a new comment on an article
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /comments/{commentId}:
    put:
      summary: Update comment
      description: Update an existing comment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdate'
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the comment author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete comment
      description: Delete a comment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Comment deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the comment author or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /connections:
    post:
      summary: Create connection
      description: Send a connection request to another user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionCreate'
      responses:
        '201':
          description: Connection request sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Connection already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /connections/{connectionId}/accept:
    post:
      summary: Accept connection
      description: Accept a connection request
      security:
        - bearerAuth: []
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connection accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the recipient of the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Connection request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        role:
          type: string
          enum: [user, moderator, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        contributionCount:
          type: integer
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
...
    CommentCreate:
      type: object
      required:
        - articleId
        - content
      properties:
        articleId:
          type: string
          format: uuid
        content:
          type: string
          minLength: 1
          maxLength: 10000
        parentId:
          type: string
          format: uuid
    CommentUpdate:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 10000
    CommentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        author:
          $ref: '#/components/schemas/UserResponse'
        articleId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isApproved:
          type: boolean
    ConnectionCreate:
      type: object
      required:
        - userId
        - connectionType
      properties:
        userId:
          type: string
          format: uuid
        connectionType:
          type: string
          enum: [friend, follow]
    ConnectionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        follower:
          $ref: '#/components/schemas/UserResponse'
        followed:
          $ref: '#/components/schemas/UserResponse'
        status:
          type: string
          enum: [pending, accepted, blocked]
        connectionType:
          type: string
          enum: [friend, follow]
        createdAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object